+++ test/elixir/fixtures/at_exit.exs
@@ -1,6 +1,6 @@
 defmodule AtExit do
   def at_exit(str) do
-    System.at_exit fn(_) -> IO.print(str) end
+    System.at_exit fn(_) -> IO.write(str) end
   end
 end
 System.at_exit fn(status) -> IO.puts "cruel world with status #{status}" end